Broke client up into several mixins
